Refactor: Page & Button Component and Share Functionality via Props
