+++ src/components/buttons/Button.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Button = ({ onClick, text, className }) => {
+  return (
+    <button className={className} onClick={onClick}>
+      {text}
+    </button>
+  );
+};
+
+export default Button;
