+++ src/components/index.js
@@ -1,3 +1,4 @@
-export { default as Listing } from './Listing';
-export { default as SearchBar } from './SearchBar';
-export { default as LoadingSpinner } from './LoadingSpinner';
+export { default as Listing } from './listing/Listing';
+export { default as SearchBar } from './SearchBar/SearchBar';
+export { default as LoadingSpinner } from './Spinner/LoadingSpinner';
+export { default as Button } from './buttons/Button';
