+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export { default as ListingPage } from './ListingPage';
-export { default as DetailsPage } from './DetailsPage';
+export { default as ListingPage } from './Listing/ListingPage';
+export { default as DetailsPage } from './details/DetailsPage';
